Once source of truth for dates, move things into the date hook
